change minBet to be double

+++ casino/src/main/java/game/Billiards.java
@@ -1,7 +1,7 @@
 package game;
 
 public class Billiards extends GameAbstract {
-    public Billiards(int maxPlayers, int minBet, double cashOutMult, double winChance, int gameTime){
+    public Billiards(int maxPlayers, double minBet, double cashOutMult, double winChance, int gameTime){
         super("Billiards", maxPlayers, minBet, cashOutMult, winChance, gameTime);
     }
 }
