+++ casino/src/main/java/game/Pachinko.java
@@ -1,7 +1,7 @@
 package game;
 
 public class Pachinko extends GameAbstract {
-    public Pachinko(int maxPlayers, int minBet, double cashOutMult, double winChance, int gameTime){
+    public Pachinko(int maxPlayers, double minBet, double cashOutMult, double winChance, int gameTime){
         super("Pachinko", maxPlayers, minBet, cashOutMult, winChance, gameTime);
     }
 }
